+++ gtfs2geojson/__init__.py
@@ -7,7 +7,7 @@ from .converter import (
     write,
 )
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"
 __all__ = [
     "convert",
     "write",